refactor : Path 클래스 domain 패키지 하위로 이동
